feat: move the genesis submodule from fixtures/networks to tools/gengen/util
