+++ cmd/daemon.go
@@ -26,6 +26,8 @@ import (
 	"github.com/filecoin-project/venus/pkg/journal"
 	"github.com/filecoin-project/venus/pkg/migration"
 	"github.com/filecoin-project/venus/pkg/repo"
+
+	gengen "github.com/filecoin-project/venus/tools/gengen/util"
 )
 
 var log = logging.Logger("daemon")
@@ -149,7 +151,7 @@ func initRun(req *cmds.Request) error {
 		genesisFunc = genesis.MakeGenesis(req.Context, rep, mkGen, preTp.(string), cfg.NetworkParams.ForkUpgradeParam)
 	} else {
 		genesisFileSource, _ := req.Options[GenesisFile].(string)
-		genesisFunc, err = networks.LoadGenesis(req.Context, rep, genesisFileSource, network)
+		genesisFunc, err = gengen.LoadGenesis(req.Context, rep, genesisFileSource, network)
 		if err != nil {
 			return err
 		}
