+++ tools/gengen/util/generator.go
@@ -46,6 +46,18 @@ import (
 	"github.com/filecoin-project/venus/pkg/vm/gas"
 	"github.com/filecoin-project/venus/pkg/vmsupport"
 	"github.com/filecoin-project/venus/venus-shared/types"
+
+	"io/ioutil"
+	"os"
+
+	"github.com/ipld/go-car"
+
+	"net/http"
+	"net/url"
+
+	"github.com/filecoin-project/venus/fixtures/assets"
+	"github.com/filecoin-project/venus/fixtures/networks"
+	"github.com/filecoin-project/venus/pkg/repo"
 )
 
 const InitialBaseFee = 100e6
@@ -783,4 +795,89 @@ func computeSectorPower(size abi.SectorSize, duration abi.ChainEpoch, dealWeight
 	spower := big.NewIntUnsigned(uint64(size))
 	qapower := miner.QAPowerForWeight(size, duration, dealWeight, verifiedDealWeight)
 	return spower, qapower
+}
+
+func LoadGenesis(ctx context.Context, rep repo.Repo, sourceName string, network string) (genesis.InitFunc, error) {
+	var (
+		source io.ReadCloser
+		err    error
+	)
+
+	if sourceName == "" {
+		networkType, err := networks.GetNetworkFromName(network)
+		if err != nil {
+			return nil, err
+		}
+
+		bs, err := assets.GetGenesis(networkType)
+		if err != nil {
+			return MakeGenesisFunc(), nil
+		}
+		source = ioutil.NopCloser(bytes.NewReader(bs))
+		// return MakeGenesisFunc(), nil
+	} else {
+		source, err = openGenesisSource(sourceName)
+		if err != nil {
+			return nil, err
+		}
+	}
+
+	defer func() { _ = source.Close() }()
+
+	genesisBlk, err := extractGenesisBlock(ctx, source, rep)
+	if err != nil {
+		return nil, err
+	}
+
+	gif := func(cst cbor.IpldStore, bs blockstore.Blockstore) (*types.BlockHeader, error) {
+		return genesisBlk, err
+	}
+
+	return gif, nil
+}
+
+func extractGenesisBlock(ctx context.Context, source io.ReadCloser, rep repo.Repo) (*types.BlockHeader, error) {
+	bs := rep.Datastore()
+	ch, err := car.LoadCar(ctx, bs, source)
+	if err != nil {
+		return nil, err
+	}
+
+	// need to check if we are being handed a car file with a single genesis block or an entire chain.
+	bsBlk, err := bs.Get(ctx, ch.Roots[0])
+	if err != nil {
+		return nil, err
+	}
+	cur, err := types.DecodeBlock(bsBlk.RawData())
+	if err != nil {
+		return nil, err
+	}
+
+	return cur, nil
+}
+
+func openGenesisSource(sourceName string) (io.ReadCloser, error) {
+	sourceURL, err := url.Parse(sourceName)
+	if err != nil {
+		return nil, fmt.Errorf("invalid filepath or URL for genesis file: %s", sourceURL)
+	}
+	var source io.ReadCloser
+	if sourceURL.Scheme == "http" || sourceURL.Scheme == "https" {
+		// NOTE: This code is temporary. It allows downloading a genesis block via HTTP(S) to be able to join a
+		// recently deployed staging devnet.
+		response, err := http.Get(sourceName)
+		if err != nil {
+			return nil, err
+		}
+		source = response.Body
+	} else if sourceURL.Scheme != "" {
+		return nil, fmt.Errorf("unsupported protocol for genesis file: %s", sourceURL.Scheme)
+	} else {
+		file, err := os.Open(sourceName)
+		if err != nil {
+			return nil, err
+		}
+		source = file
+	}
+	return source, nil
 }
